refactor: Move Store and WindowManager to models
