+++ packages/neuron-wallet/src/models/store.ts
@@ -1,6 +1,6 @@
 import EventEmitter from 'events'
 import FileService from '../services/file'
-import logger from './logger'
+import logger from '../utils/logger'
 
 class Store extends EventEmitter {
   public moduleName: string
