+++ packages/neuron-wallet/src/models/window-manager.ts
@@ -1,6 +1,6 @@
 import { BrowserWindow } from 'electron'
-import { Channel } from './const'
-import logger from './logger'
+import { Channel } from '../utils/const'
+import logger from '../utils/logger'
 
 const error = { level: 'error', message: 'Electron is not loaded' }
 
@@ -12,7 +12,7 @@ interface SendMessage {
   (channel: Channel.Chain, method: 'status' | 'tipBlockNumber', params: any): void
 }
 
-class WindowManage {
+export default class WindowManager {
   public static broadcast: SendMessage = (channel: Channel, method: string, params: any): void => {
     if (!BrowserWindow) {
       logger.log(error)
@@ -36,5 +36,3 @@ class WindowManage {
     }
   }
 }
-
-export default WindowManage
